feat: use headless renderer for Approving ChartSmart images
One source of frustration with onboarding to the exercise is
ChartSmart test failures due to pixel differences in rendering
the JPanels. This change introduces an Approval class that
uses a headless renderer to generate the images.

Whether this reduces the frustration remains to be seen, as it
needs to be tested on another machine.

Took 2 hours 17 minutes